+++ Tests/Fluxor.Blazor.Web.UnitTests/SupportFiles/FluxorComponentThatOptionallyCallsBaseOnInitialized.cs
@@ -1,0 +1,17 @@
+﻿using Fluxor.Blazor.Web.Components;
+
+namespace Fluxor.Blazor.Web.UnitTests.SupportFiles
+{
+	public class FluxorComponentThatOptionallyCallsBaseOnInitialized : FluxorComponent
+	{
+		public bool CallBaseOnInitialized;
+
+		protected override void OnInitialized()
+		{
+			if (CallBaseOnInitialized)
+				base.OnInitialized();
+		}
+
+		public void Test_OnInitialized() => OnInitialized();
+	}
+}